Added timeout; handled if user on multiple servers
So the timeout thing actually works now - it adds a Date.now() property to the server DMs and checks if it was in the last 24 hours.
As for multiple servers - what if someone dms the bot and they're on multiple servers? How can it handle mod-mail? Now it sends an embeds which asks people to choose a server. Solved!