Changed modmail for future paginator, added JSONify!
Changed the process of collating messagereactions so that it can work with changing the paginator mech for the future. Now hopefully we can have more things which have awaiting reactions and some permenant ones.
Added a JSONify command that turns an embed into its underlying JSON.
Also, fixed some bugs with eval.js